requirements.txt
pandas==2.2.0
openpyxl==3.1.2
pymongo==4.6.1
python-dotenv==1.0.0
xlrd==2.0.1